Refactoring Brain component setup
