+++ app/components/Brain/index.tsx
@@ -1,0 +1,2 @@
+import Brain from "./Brain"
+export default Brain
